CompressBatchedTask: move to pocketmine\network\mcpe namespace
